backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.34
pydantic==2.8.2
pymysql==1.1.1
redis==5.0.8
celery==5.4.0
python-dotenv==1.0.1
alembic==1.13.2
openai==1.43.0
requests==2.32.3
